refactored augmentation module to reduce the redundant code in the pipeline
